+++ src/lib/types/data/index.ts
@@ -4,3 +4,12 @@ export interface ITranslation{
      translationString: string,
      lineNumber: number
 }
+
+export interface ILangItem {
+     flag: string
+     name: string
+     code: string
+     type: "country" | "language"
+     label: string
+     value: string
+}