+++ src/popups/modals/go-to-key.tsx
@@ -6,12 +6,18 @@ import { Controller, useForm } from "react-hook-form"
 import { GoToKeyNameType } from "@/schemas/types";
 import { GoToKeyNameSchema } from "@/schemas";
 import { DialogFooter } from "@/components/ui/dialog";
-import { InputGroup, InputGroupAddon, InputGroupInput } from "@/components/ui/input-group";
-import { Search } from "lucide-react";
 import TranslatorActions from "@/actions/translator";
 import { PopupFormProps } from "@/lib/types";
+import { useAppTranslation } from "@/context/translation";
+import { useState } from "react";
+import { toast } from "sonner";
+import { getErrorMessage } from "@/lib/utils";
+import ComboboxField from "@/components/fields/combobox-field";
+import { Search } from "lucide-react";
 
 export default function GoToKeyNamePopup({triggerButton}: PopupFormProps){
+     const {table, keyNames, setCurrentTranslation, setVisibleCount, setInput} = useAppTranslation()
+     const [open, setOpen] = useState(false)
      const form = useForm<GoToKeyNameType>({
           resolver: zodResolver(GoToKeyNameSchema),
           defaultValues: {
@@ -19,7 +25,31 @@ export default function GoToKeyNamePopup({triggerButton}: PopupFormProps){
           }
      })
      const onSubmit = (values: GoToKeyNameType) => {
-          TranslatorActions.goToKeyName(values)
+          try {
+               const res = TranslatorActions.goToKeyName(values,table);
+               if(res.error) toast.error("Failed to jump into the specified key name",{
+                    description: res.error
+               })
+               if(res.success) {
+                    setCurrentTranslation(res.translation)
+                    setInput(res.translation.translationString)
+                    setVisibleCount(prev => Math.max(prev, res.index + 1))
+                    requestAnimationFrame(() => {
+                         document
+                              .getElementById(`row-${res.translation.keyName}`)
+                              ?.scrollIntoView({
+                                   behavior: "smooth",
+                                   block: "center",
+                              })
+                    })
+                    setOpen(false);
+                    form.reset()
+               }
+          } catch (err) {
+               toast.error("Failed to jump into the specified key name",{
+                    description: getErrorMessage(err)
+               })
+          }
      }
      return (
           <AppModal
@@ -27,6 +57,7 @@ export default function GoToKeyNamePopup({triggerButton}: PopupFormProps){
                title="Go to translation key name"
                description="Jump into the provided key name"
                triggerButton={triggerButton}
+               open={open} onOpenChange={setOpen}
           >
                <form id="go-to-key" onSubmit={form.handleSubmit(onSubmit)}>
                     <FieldGroup>
@@ -36,17 +67,13 @@ export default function GoToKeyNamePopup({triggerButton}: PopupFormProps){
                               render={({field, fieldState})=>(
                                    <Field data-invalid={fieldState.invalid}>
                                         <FieldLabel htmlFor={field.name}>Translation Key Name</FieldLabel>
-                                        <InputGroup>
-                                             <InputGroupInput
-                                                  {...field}
-                                                  id={field.name}
-                                                  aria-invalid={fieldState.invalid}
-                                                  placeholder="command.buttons.cancel"
-                                             />
-                                             <InputGroupAddon>
-                                                  <Search/>
-                                             </InputGroupAddon>
-                                        </InputGroup>
+                                        <ComboboxField
+                                             {...field}
+                                             invalid={fieldState.invalid}
+                                             placeholder="command.buttons.cancel"
+                                             items={keyNames}
+                                             Icon={Search}
+                                        />
                                         {fieldState.invalid && (
                                              <FieldError errors={[fieldState.error]} />
                                         )}